Rename ActivityDetailsWidget into the SurveyWidget
