+++ src/pages/PublicActivityDetails/index.tsx
@@ -2,7 +2,7 @@ import { useParams } from 'react-router-dom';
 
 import Box from '~/shared/ui/Box';
 import { useCustomTranslation } from '~/shared/utils';
-import { SurveyBasicContext, ActivityDetailsWidget } from '~/widgets/ActivityDetails';
+import { SurveyBasicContext, SurveyWidget } from '~/widgets/Survey';
 
 function PublicActivityDetailsPage() {
   const { appletId, activityId, eventId, publicAppletKey } = useParams();
@@ -17,7 +17,7 @@ function PublicActivityDetailsPage() {
       <SurveyBasicContext.Provider
         value={{ appletId, activityId, eventId, publicAppletKey, isPublic: true }}
       >
-        <ActivityDetailsWidget />
+        <SurveyWidget />
       </SurveyBasicContext.Provider>
     </Box>
   );
