--- src/widgets/ActivityDetails/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './ActivityDetails';
