+++ src/widgets/Survey/index.ts
@@ -1,2 +1,2 @@
-export * from './ui';
+export { SurveyWidget } from './ui';
 export { SurveyBasicContext } from './lib';
