+++ src/widgets/Survey/lib/index.ts
