+++ src/widgets/Survey/model/generateUserPublicKey.ts
