+++ src/widgets/Survey/model/getFirstResponseDataIdentifierTextItem.ts
