+++ src/widgets/Survey/model/getScheduledTimeFromEvents.ts
