+++ src/widgets/Survey/model/hooks/useActivityDetailsQuery.ts
