+++ src/widgets/Survey/model/hooks/useAutoForward.ts
