+++ src/widgets/Survey/model/index.ts
