+++ src/widgets/Survey/model/mappers.ts
