+++ src/widgets/Survey/model/prepareItemAnswers.ts
