+++ src/widgets/Survey/model/validateItem.ts
