+++ src/widgets/Survey/ui/ActivityMetaData.tsx
