+++ src/widgets/Survey/ui/AssessmentLayout.tsx
