+++ src/widgets/Survey/ui/AssessmentLayoutFooter.tsx
