+++ src/widgets/Survey/ui/AssessmentLayoutHeader.tsx
