+++ src/widgets/Survey/ui/AssessmentLoadingScreen.tsx
