+++ src/widgets/Survey/ui/AssessmentPassingScreen.tsx
