+++ src/widgets/Survey/ui/AssessmentSummaryScreen.tsx
