+++ src/widgets/Survey/ui/AssessmentWelcomeScreen.tsx
